Adds test components.

+++ src/components.rs
@@ -4,19 +4,19 @@ use specs::{Component, DenseVecStorage};
 #[derive(Component)]
 pub struct Transform {
     pub position: Vector2,
-    pub rotation: f64,
+    pub rotation: f32,
 }
 
 #[derive(Component)]
 pub struct RigidBody {
     pub velocity: Vector2,
-    pub spin: f64,
+    pub spin: f32,
 }
 
 #[derive(Component)]
 pub struct Collider {
     pub collider_type: ColliderType,
-    pub radius: f64,
+    pub radius: f32,
     pub size: Vector2,
     pub is_trigger: bool,
 }
